requirements.txt
openai>=1.60.0
pydantic>=2.8.0
python-dateutil>=2.9.0
python-dotenv>=1.0.1
pypdf>=5.0.0
langdetect>=1.0.9
pytest>=8.2.0
